hf-space/requirements.txt
flask==3.0.0
flask-cors==4.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
tensorflow>=2.15.0
numpy>=1.24.3
Pillow>=10.1.0
werkzeug>=3.0.1
scikit-learn>=1.3.2
matplotlib>=3.7.2
requests>=2.31.0
keras-hub>=0.25.0
